Attempt to make properties with physical meaning

+++ DeviceAdapters/MicroFPGA/MicroFPGA.h
@@ -104,10 +104,10 @@ private:
 	bool initialized_;
 	bool mode_;
 	bool start_;
-	long pulse_;
-	long period_;
-	long exposure_;
-	long delay_;
+	double pulse_;
+	double period_;
+	double exposure_;
+	double delay_;
 	bool busy_;
 };
